Add real development support
Reworked Dockerfile to not create a sample application.

Added `create-new-app.sh` script that creates a sample project and stores it in the same directory at `./app`.

Starting the Docker Compose `app` service bind mounts this directory and resources can be edited in real time and are reflected in the application almost immediately via Webpack.