Use environment variables for database configuration
Update `create-new-db-app.sh` so that it reconfigures `config/dev.exs` to use environment variables for configuring the database connection. These are written to `.env` and pulled in via `docker-compose.yml`.

Also patched the `create-new-db-app.sh` script so that it actually creates the applications. Was commented out for debugging purposes and not re-enabled on the last push.

Update `README.md` to also point to the [secrets section](https://hexdocs.pm/phoenix/deployment.html#handling-of-your-application-secrets) of the Phoenix documentation.